src: Export toString to Qt5 XML
